The portfolio of HanHanZheng
The designs in my portfolio enhance both my design and artistic ability to a large extent.  In daily life, I focus on subtleties to seize inspiration: When I paid visitation in West Lake,  many service facilities are found scattered throughout the landscape. So I made a  design of the west lake scenic station that provides relaxation, shopping, purchasing tickets, and a special platform for live streaming. the shape of the post was designed like a lotus to highlight the beauty of the classical rhythms characteristic of the Zhejiang region. The process of designing made me aware of the importance of practicability, to consider more about the aspects of fields. However, my practical attitude was challenged when I create a conceptual model for providing wild cat services. I tried to stream more personal ideas and concepts in completion, hence augmenting the contact between cats and passers-by. Furthermore, the painting of a sketchbook is the most rewarding program in my portfolio. With the help of my teacher, I enlarge my knowledge of painting comprehensively, such as using different materials to form a painting rather than painting with lines or pigments. The increase in painting ability enables me to create more boldly with a variety of materials and has opened up more possibilities for my designs.